Simplify testing approach - focus on unit tests and basic smoke tests
- Removed complex Docker-resilient integration tests that tested error handling more than functionality
- Created pure unit tests for BaaS detection using dependency injection instead of file system operations
- Replaced integration tests with simple smoke tests that validate CLI basics (help, version, init)
- Updated Vitest config to use process forks to support process.chdir in smoke tests
- Streamlined GitHub Actions workflow to run unit tests + smoke tests only

This approach focuses on testing what matters: pure functions and basic CLI functionality, rather than Docker orchestration in Docker-less environments.
