+++ vitest.config.ts
@@ -4,6 +4,12 @@ export default defineConfig({
   test: {
     globals: true,
     environment: 'node',
+    pool: 'forks', // Use process forks instead of workers to support process.chdir
+    poolOptions: {
+      forks: {
+        singleFork: true, // Use single fork for smoke tests that need process.chdir
+      },
+    },
     coverage: {
       provider: 'v8',
       reporter: ['text', 'json', 'html'],
@@ -16,7 +22,7 @@ export default defineConfig({
         'docs/',
       ],
     },
-    include: ['tests/**/*.test.ts'],
+    include: ['tests/unit/**/*.test.ts', 'tests/smoke/**/*.test.ts'],
     testTimeout: 10000,
     hookTimeout: 10000,
   },
